Mana issues

+++ Assets/Systems/Deletion System.cs
@@ -10,7 +10,8 @@ public class DeletionSystem : ComponentSystem
         
         Entities.ForEach((Entity e,ref DeleteComp d) =>
         {
-            World.Active.EntityManager.DestroyEntity(e);
+            if (d.buffer == 0) World.Active.EntityManager.DestroyEntity(e);
+            else d.buffer--;
         });
     }
 }
